feat: only allow users to supply an account based xpub for creating watch-only wallets
This is due to a quirk in LibWally-Swift where single sig PSBT's do not get successfully signed by account based xprv's. We can expand on this functionality later but for now to keep the UI and core functionality consistent along with a clean codebase the user can simply supply a BIP39 mnemonic if they want to supply there own seed.