feat: abilty to recover accounts with ur:crypto-seed
Just scan a QR of a ur:crypto-seed from the accounts tab or paste one in to the "recover with words" text field and tap add or done, the app will then parse the ur and convert it to BIP39 words if it can. For now the birthdate field is ignored as it presents some issues.