Alert bug fix for iPad and update Account Map format
A recent iOS update broke the way UIAlerts would display on iPad, this commit fixes that so that action alerts should now display properly. After discussions with the Specter team, we amended the Account Map format slightly so that the checksum is not appended, and the single quotes are replaced with `h` to denote hardened path components, the reason being single quotes can cause formatting issues. It is now up to the client to recalculate the checksum with bitcoin-cli getdescriptorinfo and store the Bitcoin Core compatible descriptor for local use.